@@ -6,7 +6,7 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.2.6.RELEASE</version>
+    <version>2.5.8</version>
     <relativePath /> <!-- lookup parent from repository -->
   </parent>
   <groupId>com.google.cloud.healthcare.fdamystudies</groupId>
@@ -17,9 +17,9 @@
   <description>Common service module for FDA MyStudies</description>
 
   <properties>
-    <java.version>1.8</java.version>
+    <java.version>17</java.version>
     <common.text.version>1.8</common.text.version>
-    <mail.version>1.4</mail.version>
+    <mail.version>1.6.2</mail.version>
   </properties>
   <dependencyManagement>
     <dependencies>
@@ -52,6 +52,10 @@
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-validation</artifactId>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
@@ -90,7 +94,7 @@
     </dependency>
     <dependency>
       <groupId>javax.mail</groupId>
-      <artifactId>mail</artifactId>
+      <artifactId>javax.mail-api</artifactId>
       <version>${mail.version}</version>
     </dependency>
     <dependency>
@@ -175,7 +179,7 @@
       <plugin>
         <groupId>com.google.cloud.tools</groupId>
         <artifactId>jib-maven-plugin</artifactId>
-        <version>2.5.2</version>
+        <version>3.4.1</version>
         <configuration>
           <!-- we don't want jib to execute on this module -->
           <skip>true</skip>
@@ -184,9 +188,10 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
+        <version>3.12.1</version>
         <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
+          <source>17</source>
+          <target>17</target>
           <fork>true</fork>
           <compilerArgs>
             <arg>-J--add-opens=jdk.compiler/com.sun.tools.javac.code=ALL-UNNAMED</arg>
@@ -204,7 +209,7 @@
             <path>
               <groupId>org.projectlombok</groupId>
               <artifactId>lombok</artifactId>
-              <version>1.18.12</version>
+              <version>1.18.30</version>
             </path>
           </annotationProcessorPaths>
         </configuration>
