@@ -6,7 +6,7 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.2.6.RELEASE</version>
+    <version>2.5.8</version>
     <relativePath /> <!-- lookup parent from repository -->
   </parent>
   <groupId>com.google.cloud.healthcare.fdamystudies</groupId>
@@ -17,7 +17,7 @@
   <description>Common tests for FDA MyStudies</description>
 
   <properties>
-    <java.version>1.8</java.version>
+    <java.version>17</java.version>
     <wiremock.version>2.26.3</wiremock.version>
     <common.text.version>1.8</common.text.version>
   </properties>
@@ -105,7 +105,7 @@
       <plugin>
         <groupId>com.google.cloud.tools</groupId>
         <artifactId>jib-maven-plugin</artifactId>
-        <version>2.5.2</version>
+        <version>3.4.1</version>
         <configuration>
           <!-- we don't want jib to execute on this module -->
           <skip>true</skip>
@@ -114,9 +114,10 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
+        <version>3.12.1</version>
         <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
+          <source>17</source>
+          <target>17</target>
           <fork>true</fork>
           <compilerArgs>
             <arg>-J--add-opens=jdk.compiler/com.sun.tools.javac.code=ALL-UNNAMED</arg>
@@ -134,7 +135,7 @@
             <path>
               <groupId>org.projectlombok</groupId>
               <artifactId>lombok</artifactId>
-              <version>1.18.12</version>
+              <version>1.18.30</version>
             </path>
           </annotationProcessorPaths>
         </configuration>
