@@ -6,7 +6,7 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.2.6.RELEASE</version>
+    <version>2.5.8</version>
     <relativePath/> <!-- lookup parent from repository -->
   </parent>
   <groupId>com.google.cloud.healthcare.fdamystudies</groupId>
@@ -17,11 +17,11 @@
   <description>Project containing user management services</description>
 
 	<properties>
-		<java.version>1.8</java.version>
+		<java.version>17</java.version>
 		<junit-jupiter.version>5.5.2</junit-jupiter.version>
 		<wiremock.version>2.26.3</wiremock.version>
 		<clover.version>4.4.1</clover.version>
-		<mockito.version>2.24.0</mockito.version>
+		<mockito.version>3.4.6</mockito.version>
 	</properties>
 
 	<dependencies>
@@ -276,10 +276,10 @@
       <plugin>
         <groupId>com.google.cloud.tools</groupId>
         <artifactId>jib-maven-plugin</artifactId>
-        <version>2.6.0</version>
+        <version>3.4.1</version>
         <configuration>
           <from>
-            <image>tomcat:9.0.33-jdk11-openjdk</image>
+            <image>tomcat:9.0-jdk17-openjdk</image>
           </from>
           <container>
             <appRoot>/usr/local/tomcat/webapps/participant-consent-datastore</appRoot>
@@ -289,9 +289,10 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
-        <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
+        <version>3.12.1</version>
+              <configuration>
+          <source>17</source>
+          <target>17</target>
           <fork>true</fork>
           <compilerArgs>
             <arg>-J--add-opens=jdk.compiler/com.sun.tools.javac.code=ALL-UNNAMED</arg>
@@ -309,7 +310,7 @@
             <path>
               <groupId>org.projectlombok</groupId>
               <artifactId>lombok</artifactId>
-              <version>1.18.12</version>
+              <version>1.18.30</version>
             </path>
           </annotationProcessorPaths>
         </configuration>
