@@ -8,7 +8,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.2.6.RELEASE</version>
+		<version>2.5.8</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.google.cloud.healthcare.fdamystudies</groupId>
@@ -19,7 +19,7 @@
 	<description>Response Server web service components for the FDA MyStudies Open Implementation</description>
 
 	<properties>
-		<java.version>1.8</java.version>
+		<java.version>17</java.version>
 		<clover.version>4.4.1</clover.version>
 		<common.text.version>1.8</common.text.version>
 		<wiremock.version>2.26.3</wiremock.version>
@@ -68,10 +68,10 @@
 			<plugin>
 				<groupId>com.google.cloud.tools</groupId>
 				<artifactId>jib-maven-plugin</artifactId>
-				<version>2.6.0</version>
+				<version>3.4.1</version>
 				<configuration>
 					<from>
-						<image>tomcat:9.0.33-jdk11-openjdk</image>
+						<image>tomcat:9.0-jdk17-openjdk</image>
 					</from>
 					<container>
 						<appRoot>/usr/local/tomcat/webapps/response-datastore</appRoot>
@@ -81,9 +81,10 @@
 			<plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
+        <version>3.12.1</version>
         <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
+          <source>17</source>
+          <target>17</target>
           <fork>true</fork>
           <compilerArgs>
             <arg>-J--add-opens=jdk.compiler/com.sun.tools.javac.code=ALL-UNNAMED</arg>
@@ -101,7 +102,7 @@
             <path>
               <groupId>org.projectlombok</groupId>
               <artifactId>lombok</artifactId>
-              <version>1.18.12</version>
+              <version>1.18.30</version>
             </path>
           </annotationProcessorPaths>
         </configuration>
