@@ -12,8 +12,8 @@
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <app.dir.name>fdahpStudyDesigner</app.dir.name>
-    <org.springframework.version>5.3.18</org.springframework.version>
-    <spring.security.version>5.6.2</spring.security.version>
+    <org.springframework.version>5.3.32</org.springframework.version>
+    <spring.security.version>5.8.9</spring.security.version>
   </properties>
   <build>
     <finalName>studybuilder</finalName>
@@ -32,12 +32,16 @@
         </configuration>
       </plugin>
       <plugin>
+		<artifactId>maven-war-plugin</artifactId>
+		<version>3.3.1</version>
+	  </plugin>
+      <plugin>
         <artifactId>maven-compiler-plugin</artifactId>
-        <version>2.3.2</version>
+        <version>3.12.1</version>
         <configuration>
           <webXml>WEB-INF/web.xml</webXml>
-          <source>1.7</source>
-          <target>1.7</target>
+          <source>17</source>
+          <target>17</target>
         </configuration>
       </plugin>
       <plugin>
@@ -61,9 +65,9 @@
       <plugin>
         <groupId>com.google.cloud.tools</groupId>
         <artifactId>jib-maven-plugin</artifactId>
-        <version>2.6.0</version>
+        <version>3.4.1</version>
         <configuration>
-          <from><image>tomcat:9.0.33-jdk11-openjdk</image></from>
+          <from><image>tomcat:9.0-jdk17-openjdk</image></from>
           <container><appRoot>/usr/local/tomcat/webapps/studybuilder</appRoot></container>
         </configuration>
       </plugin>
@@ -332,7 +336,7 @@
     <dependency>
       <groupId>org.hibernate</groupId>
       <artifactId>hibernate-core</artifactId>
-      <version>5.6.7.Final</version>
+      <version>5.6.15.Final</version>
       <type>jar</type>
       <scope>compile</scope>
     </dependency>
@@ -563,7 +567,7 @@
     <dependency>
       <groupId>org.projectlombok</groupId>
       <artifactId>lombok</artifactId>
-      <version>1.18.12</version>
+      <version>1.18.30</version>
     </dependency>
     <dependency>
       <groupId>org.slf4j</groupId>
