@@ -92,12 +92,12 @@
 	<dependency>
 	    <groupId>org.glassfish.jersey.containers</groupId>
 	    <artifactId>jersey-container-servlet</artifactId>
-	    <version>2.6</version>
+	    <version>2.35</version>
 	</dependency>
 	<dependency>
 	    <groupId>org.glassfish.jersey.bundles</groupId>
 	    <artifactId>jaxrs-ri</artifactId>
-	    <version>2.6</version>
+	    <version>2.35</version>
 	</dependency>
 
 		<dependency>
@@ -155,20 +155,20 @@
 			<artifactId>mysql-connector-java</artifactId>
 			<version>8.0.16</version>
 		</dependency>
-		<dependency>
+		<!--<dependency>
 			<groupId>org.hibernate</groupId>
 			<artifactId>hibernate-annotations</artifactId>
 			<version>3.5.5-Final</version>
-		</dependency>
+		</dependency>-->
 		<dependency>
 			<groupId>org.hibernate</groupId>
 			<artifactId>hibernate-core</artifactId>
-			<version>3.5.5-Final</version>
+			<version>5.6.15.Final</version>
 		</dependency>
 		<dependency>
 			<groupId>org.hibernate</groupId>
 			<artifactId>hibernate-entitymanager</artifactId>
-			<version>3.5.5-Final</version>
+			<version>5.6.15.Final</version>
 		</dependency>
 
 		<dependency>
@@ -285,11 +285,11 @@
 			</plugin>
 			<plugin>
 				<artifactId>maven-compiler-plugin</artifactId>
-				<version>2.3.2</version>
+				<version>3.12.1</version>
 				<configuration>
 					<webXml>WEB-INF/web.xml</webXml>
-					<source>1.7</source>
-					<target>1.7</target>
+					<source>17</source>
+					<target>17</target>
 				<fork>true</fork>
       <compilerArgs>
         <arg>-J--add-opens=jdk.compiler/com.sun.tools.javac.code=ALL-UNNAMED</arg>
@@ -307,7 +307,7 @@
         <path>
           <groupId>org.projectlombok</groupId>
           <artifactId>lombok</artifactId>
-          <version>1.18.12</version>
+          <version>1.18.30</version>
         </path>
       </annotationProcessorPaths>
     </configuration>
@@ -315,10 +315,10 @@
 			<plugin>
 				<groupId>com.google.cloud.tools</groupId>
 				<artifactId>jib-maven-plugin</artifactId>
-				<version>2.6.0</version>
+				<version>3.4.1</version>
 				<configuration>
 					<from>
-						<image>tomcat:9.0.33-jdk11-openjdk</image>
+						<image>tomcat:9.0-jdk17-openjdk</image>
 					</from>
 					<container>
 						<appRoot>/usr/local/tomcat/webapps/study-datastore</appRoot>
