@@ -32,9 +32,9 @@ import static com.google.cloud.healthcare.fdamystudies.utils.Constants.SUCCESS;
 import static com.google.cloud.healthcare.fdamystudies.utils.Constants.USER_ID_HEADER;
 import static com.google.cloud.healthcare.fdamystudies.utils.Constants.VALID_USER_ID;
 import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.doNothing;
